feat(auth):redirect to recent workspace if present
this change rewrites the api hooks and introduces a `refetch` function
to be called by the user of the hook to re-request data from the server
and asynchronously act on it once ready.

we need to be able to refetch data on signin, as otherwise of course the
redirect to the recent workspace is not working for we are not yet
authenticated.

the plan is to have this pattern for all api hooks if we need to
re-request the endpoint again.

also the commit introduces the fetch from the workspace endpoint. for
now we'll simply want to display an overview of pages etc.

in the future a user will be able to choose whether he wants the
workspace to open directly in a certain page, his most recently accessed
page, or the overview.